moving indrel static pages into base app
