--- hknweb/indrel/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class IndrelConfig(AppConfig):
-    name = "indrel"
