+++ hknweb/settings/common.py
@@ -32,7 +32,6 @@ INSTALLED_APPS = [
     "hknweb.candidate",
     "hknweb.coursesemester",
     "hknweb.course_surveys",
-    "hknweb.indrel",
     "hknweb.serv",
     "hknweb.studentservices",
     "hknweb.thread",
