+++ hknweb/templates/base.html
@@ -113,15 +113,15 @@
       </label>
     </div>
     <div id="industry-nav-bar" class="sub-nav-bar navi-container">
-      <a href="{% url 'indrel:index' %}">Why HKN?</a>
+      <a href="{% url 'indrel' %}">Why HKN?</a>
       <span class="slash">/</span>
-      <a href="{% url 'indrel:resume_book' %}">Resume Book</a>
+      <a href="{% url 'resume_book' %}">Resume Book</a>
       <span class="slash">/</span>
-      <a href="{% url 'indrel:infosessions' %}">Infosessions</a>
+      <a href="{% url 'infosessions' %}">Infosessions</a>
       <span class="slash">/</span>
-      <a href="{% url 'indrel:career_fair' %}">Career Fair</a>
+      <a href="{% url 'career_fair' %}">Career Fair</a>
       <span class="slash">/</span>
-      <a href="{% url 'indrel:contact_us' %}">Contact Us</a>
+      <a href="{% url 'contact_us' %}">Contact Us</a>
     </div>
     <div class="navi-container mobile-nav-bar" style="text-align: center;">
       <label for="toggle-outreach">
