+++ hknweb/templates/indrel/career_fair.html
@@ -41,6 +41,6 @@ Because we only target students from technical/engineering majors, your company 
 
 <h2>Questions?</h2>
 
-Feel free to <a href="{% url 'indrel:contact_us' %}">contact us</a> with any questions or comments you may have.
+Feel free to <a href="{% url 'contact_us' %}">contact us</a> with any questions or comments you may have.
 </div>
 {% endblock %}
