+++ hknweb/templates/indrel/infosessions.html
@@ -59,6 +59,6 @@ Yes, we can work with other organizations to coordinate and setup your infosessi
 
 <h2>Sounds good, who should I contact?</h2>
 
-If you would like to register an infosession, please feel free to <a href="{% url 'indrel:contact_us' %}">contact us</a>.
+If you would like to register an infosession, please feel free to <a href="{% url 'contact_us' %}">contact us</a>.
 </div>
 {% endblock %}
