+++ hknweb/templates/indrel/resume_book.html
@@ -42,6 +42,6 @@ Check back later to see statistics!
 
 <h2>How do I order it?</h2>
 
-If you would like to order a resume book, please feel free to <a href="{% url 'indrel:contact_us' %}">contact us</a>.
+If you would like to order a resume book, please feel free to <a href="{% url 'contact_us' %}">contact us</a>.
 </div>
 {% endblock %}
